refactor!(core): update the type of `n_predict` to `i32`

+++ llama-chat/src/main.rs
@@ -23,9 +23,9 @@ struct Cli {
     /// Size of the prompt context
     #[arg(short, long, default_value = "512")]
     ctx_size: u64,
-    /// Number of tokens to predict
-    #[arg(short, long, default_value = "1024")]
-    n_predict: u64,
+    /// Number of tokens to predict, -1 = infinity, -2 = until context filled.
+    #[arg(short, long, default_value = "-1")]
+    n_predict: i32,
     /// Number of layers to run on the GPU
     #[arg(short = 'g', long, default_value = "100")]
     n_gpu_layers: u64,
